requirements.txt
Flask
python-decouple
requests
json2html